Fix pages in App router
