+++ app/layout.js
@@ -1,12 +1,11 @@
 "use client";
 
 import Head from "next/head";
-import Image from "next/image";
 import Script from "next/script";
 
+import "../styles/globals.css";
 import styles from "./layout.module.css";
-import utilStyles from "../styles/utils.module.css";
-import Link from "next/link";
+import Providers from "./providers";
 
 const name = "NUS Aerobics";
 export const siteTitle = "Next.js Sample Website";
@@ -15,35 +14,36 @@ export default function Layout({ children, home }) {
   return (
     <html>
       <body>
-        <div className={styles.container}>
-          <Head>
-            <link rel="icon" href="/favicon.ico" />
-            <meta
-              name="description"
-              content="Learn how to build a personal website using Next.js"
+        <Providers>
+          <div className={styles.container}>
+            <Head>
+              <link rel="icon" href="/favicon.ico" />
+              <meta
+                name="description"
+                content="Learn how to build a personal website using Next.js"
+              />
+              <meta
+                property="og:image"
+                content={`https://og-image.vercel.app/${encodeURI(
+                  siteTitle
+                )}.png?theme=light&md=0&fontSize=75px&images=https%3A%2F%2Fassets.zeit.co%2Fimage%2Fupload%2Ffront%2Fassets%2Fdesign%2Fnextjs-black-logo.svg`}
+              />
+              <meta name="og:title" content={siteTitle} />
+              <meta name="twitter:card" content="summary_large_image" />
+            </Head>
+            <Script
+              src="https://connect.facebook.net/en_US/sdk.js"
+              strategy="lazyOnload"
+              onLoad={() =>
+                console.log(
+                  `script loaded correctly, window.FB has been populated`
+                )
+              }
             />
-            <meta
-              property="og:image"
-              content={`https://og-image.vercel.app/${encodeURI(
-                siteTitle
-              )}.png?theme=light&md=0&fontSize=75px&images=https%3A%2F%2Fassets.zeit.co%2Fimage%2Fupload%2Ffront%2Fassets%2Fdesign%2Fnextjs-black-logo.svg`}
-            />
-            <meta name="og:title" content={siteTitle} />
-            <meta name="twitter:card" content="summary_large_image" />
-          </Head>
-          <Script
-            src="https://connect.facebook.net/en_US/sdk.js"
-            strategy="lazyOnload"
-            onLoad={() =>
-              console.log(
-                `script loaded correctly, window.FB has been populated`
-              )
-            }
-          />
-          <header className={styles.header}>
-            {home ? (
-              <>
-                <Image
+            <header className={styles.header}>
+              {home ? (
+                <>
+                  {/* <Image
                   priority
                   src="/images/profile.jpg"
                   className={utilStyles.borderCircle}
@@ -51,11 +51,11 @@ export default function Layout({ children, home }) {
                   width={144}
                   alt={name}
                 />
-                <h1 className={utilStyles.heading2Xl}>{name}</h1>
-              </>
-            ) : (
-              <>
-                <Link href="/">
+                <h1 className={utilStyles.heading2Xl}>{name}</h1> */}
+                </>
+              ) : (
+                <>
+                  {/* <Link href="/">
                   <Image
                     priority
                     src="/images/profile.jpg"
@@ -66,20 +66,19 @@ export default function Layout({ children, home }) {
                   />
                 </Link>
                 <h2 className={utilStyles.headingLg}>
-                  <Link href="/" className={utilStyles.colorInherit}>
-                    {name}
-                  </Link>
-                </h2>
-              </>
-            )}
-          </header>
-          <main>{children}</main>
-          {!home && (
+                  <Link href="/" className={utilStyles.colorInherit}></Link>
+                </h2> */}
+                </>
+              )}
+            </header>
+            <main>{children}</main>
+            {/* {!home && (
             <div className={styles.backToHome}>
               <Link href="/">← Back to home</Link>
             </div>
-          )}
-        </div>
+          )} */}
+          </div>
+        </Providers>
       </body>
     </html>
   );
