+++ app/layout.module.css
@@ -1,8 +1,8 @@
 .container {
-  max-width: 36rem;
+  /* max-width: 36rem;
   padding: 0 1rem;
   margin: 3rem auto 6rem;
-  background-color: red;
+  background-color: red; */
 }
 
 .header {
