+++ app/providers.js
@@ -1,0 +1,6 @@
+"use client";
+
+import { ChakraProvider } from "@chakra-ui/react";
+export default function Providers({ children }) {
+  return <ChakraProvider>{children}</ChakraProvider>;
+}
